strainer::providers
Trait Provider

pub trait Provider:
    Debug
    + Any
    + Send
    + Sync {
    // Required methods
    fn get_rate_limits(&self) -> Result<RateLimitInfo>;
    fn as_any(&self) -> &dyn Any;
}

Trait defining the interface for rate limit providers

Required Methods

fn get_rate_limits(&self) -> Result<RateLimitInfo>

Get the current rate limits from the provider Get the current rate limits for the API provider

Errors

Returns an error if:

Unable to fetch rate limit information from the API

Network connectivity issues

Invalid API response format

fn as_any(&self) -> &dyn Any